# tenjin/api2/openapi.json
paths:
  /api2/projects/{project_id}:
    get:
      parameters:
      - in: path
        name: id
        description: The identifier
        schema:
          type: string
        required: true
      responses:
        '200':
          description: A project object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      parameters:
      - in: path
        name: id
        description: The identifier
        schema:
          type: string
        required: true
      responses:
        '200':
          description: A project object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdInput'
  /api2/projects:
    get:
      parameters:
      - in: path
        name: id
        description: List of IDs
        schema:
          type: array
          items:
            type: string
        required: true
        explode: true
        style: form
      - in: path
        name: shortid
        description: List of short IDs
        schema:
          type: array
          items:
            type: string
        required: true
        explode: true
        style: form
      - in: path
        name: name
        description: List of names
        schema:
          type: array
          items:
            type: string
        required: true
        explode: true
        style: form
      - in: path
        name: category_name
        description: Category name
        schema:
          type: string
        required: true
      - in: path
        name: category_id
        description: Category ID
        schema:
          type: string
        required: true
      - in: path
        name: parent_group_id
        description: Parent group ID
        schema:
          type: string
        required: true
      - in: path
        name: fielddata
        description: Field data in JSON format
        schema:
          type: array
          items:
            writeOnly: true
        required: true
        explode: true
        style: form
      - in: path
        name: include_linked
        description: Include linked groups
        schema:
          type: integer
        required: true
      responses:
        '200':
          description: A project object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectAll'
    post:
      parameters:
      - in: body
        required: false
        name: body
        schema:
          $ref: '#/components/schemas/ProjectPost'
      responses:
        '200':
          description: A project object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdInput'
info:
  title: FURTHRmind REST API
  version: 1.0.0
openapi: 3.0.2
components:
  schemas:
    InnerSchemaWithShortID:
      type: object
      properties:
        id:
          type: string
          description: The identifier
        name:
          type: string
          description: The name
        shortID:
          type: string
          description: The short identifier
      additionalProperties: false
    InnerUnit:
      type: object
      properties:
        id:
          type: string
          description: The unit identifier
        name:
          type: string
          description: Unit short name
        longname:
          type: string
          description: Unit long name
      additionalProperties: false
    Owner:
      type: object
      properties:
        id:
          type: string
          description: The user identifier
        email:
          type: string
          description: The user's email address
      required:
      - email
      - id
      additionalProperties: false
    User:
      type: object
      properties:
        id:
          type: string
          description: The user identifier
        email:
          type: string
          description: The user's email address
        read:
          type: boolean
          description: Has read access
        write:
          type: boolean
          description: Has write access
        delete:
          type: boolean
          description: Has delete access
        admin:
          type: boolean
          description: Has admin rights
        owner:
          type: boolean
          description: User is project owner
      required:
      - admin
      - delete
      - email
      - id
      - owner
      - read
      - write
      additionalProperties: false
    UserGroup:
      type: object
      properties:
        name:
          type: string
          description: The name of the user group
        read:
          type: boolean
          description: Has read access
        write:
          type: boolean
          description: Has write access
        delete:
          type: boolean
          description: Has delete access
        admin:
          type: boolean
          description: Has admin access
      required:
      - admin
      - delete
      - name
      - read
      - write
      additionalProperties: false
    Permission:
      type: object
      properties:
        owner:
          description: The owner of the project
          allOf:
          - $ref: '#/components/schemas/Owner'
        users:
          type: array
          description: User with permission to the project
          items:
            $ref: '#/components/schemas/User'
        usergroups:
          type: array
          description: User groups with permission to the
          items:
            $ref: '#/components/schemas/UserGroup'
      additionalProperties: false
    InnerSchemaWithoutShortID:
      type: object
      properties:
        id:
          type: string
          description: The identifier
        name:
          type: string
          description: The name
      required:
      - id
      - name
      additionalProperties: false
    InnerField:
      type: object
      properties:
        id:
          type: string
          description: The field identifier
        name:
          type: string
          description: The field name
        type:
          type: string
          description: The field type
        comboboxentries:
          type: array
          description: The field combobox entries
          items:
            $ref: '#/components/schemas/InnerSchemaWithoutShortID'
      required:
      - id
      - name
      - type
      additionalProperties: false
    Project:
      type: object
      properties:
        id:
          type: string
          description: The project identifier
        name:
          type: string
          description: Project name
        info:
          type: string
          description: Project information
        shortID:
          type: string
          description: The project short identifier
        samples:
          type: array
          description: All groups belonging to the project
          items:
            $ref: '#/components/schemas/InnerSchemaWithShortID'
        experiments:
          type: array
          description: All groups belonging to the project
          items:
            $ref: '#/components/schemas/InnerSchemaWithShortID'
        groups:
          type: array
          description: All groups belonging to the project
          items:
            $ref: '#/components/schemas/InnerSchemaWithShortID'
        units:
          type: array
          description: All units belonging to the project
          items:
            $ref: '#/components/schemas/InnerUnit'
        researchitems:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/InnerSchemaWithShortID'
        permissions:
          description: All permissions belonging to the project
          allOf:
          - $ref: '#/components/schemas/Permission'
        fields:
          type: array
          description: All field belonging to the project
          items:
            $ref: '#/components/schemas/InnerField'
      required:
      - experiments
      - fields
      - groups
      - id
      - info
      - name
      - permissions
      - samples
      - shortID
      - units
      additionalProperties: false
    ProjectAll:
      type: object
      properties:
        id:
          type: string
          description: Project identifier
        name:
          type: string
          description: Project name
        info:
          type: string
          description: Project information
      additionalProperties: false
    ProjectPost:
      type: object
      properties:
        name:
          type: string
          description: Project name
        info:
          type: string
          description: Project information
        id:
          type: string
          description: Project ID
      additionalProperties: false
    IdInput:
      type: object
      properties:
        id:
          type: string
          description: The identifier
      required:
      - id
      additionalProperties: false